move: 메인 페이지 폴더 이동 (#11)
