+++ src/app/(index)/store-list/store-list.tsx
