+++ src/app/(index)/store-list/styles.css.ts
